# feat(r): Optimize conversion from sfc to ArrowArray