--- a/r/geoarrow/R/sf-compat.R
+++ b/r/geoarrow/R/sf-compat.R
@@ -37,5 +37,5 @@ as_geoarrow_array.sfc <- function(x, ..., schema = NULL) {
 #' @importFrom nanoarrow as_nanoarrow_array
 #' @export
 as_nanoarrow_array.sfc <- function(x, ..., schema = NULL) {
-  as_geoarrow_array.sfc(x, ..., schema = schema)
+  as_geoarrow_array(x, ..., schema = schema)
 }