--- a/r/geoarrow/src/r-init.c
+++ b/r/geoarrow/src/r-init.c
@@ -13,6 +13,7 @@ SEXP geoarrow_c_kernel(SEXP kernel_name_sexp, SEXP arg_types_sexp, SEXP options_
                        SEXP schema_out_xptr);
 SEXP geoarrow_c_kernel_push(SEXP kernel_xptr, SEXP args_sexp, SEXP array_out_xptr);
 SEXP geoarrow_c_kernel_finish(SEXP kernel_xptr, SEXP array_out_xptr);
+SEXP geoarrow_c_as_nanoarrow_array_sfc(SEXP sfc, SEXP schema_xptr, SEXP array_xptr);
 
 static const R_CallMethodDef CallEntries[] = {
     {"geoarrow_c_handle_stream", (DL_FUNC)&geoarrow_c_handle_stream, 2},
@@ -23,6 +24,7 @@ static const R_CallMethodDef CallEntries[] = {
     {"geoarrow_c_kernel", (DL_FUNC)&geoarrow_c_kernel, 4},
     {"geoarrow_c_kernel_push", (DL_FUNC)&geoarrow_c_kernel_push, 3},
     {"geoarrow_c_kernel_finish", (DL_FUNC)&geoarrow_c_kernel_finish, 2},
+    {"geoarrow_c_as_nanoarrow_array_sfc", (DL_FUNC)&geoarrow_c_as_nanoarrow_array_sfc, 3},
     {NULL, NULL, 0}};
 
 void R_init_geoarrow(DllInfo* dll) {